--- a/iroh-net/bench/src/s2n.rs
+++ b/iroh-net/bench/src/s2n.rs
@@ -1,0 +1,5 @@
+use clap::Parser;
+
+#[derive(Parser, Debug, Clone, Copy)]
+#[clap(name = "s2n")]
+pub struct Opt {}